Added framework for DeleteLocalVolume gRPC call

+++ controllers/replication-controller/persistentvolume_controller.go
@@ -131,11 +131,18 @@ func (r *PersistentVolumeReconciler) Reconcile(ctx context.Context, req ctrl.Req
 					if _, ok := remoteVolume.Annotations[controller.DeletionRequested]; !ok {
 						// Add annotation on the remote PV to request its deletion
 						remoteVolumeCopy := remoteVolume.DeepCopy()
+						log.V(common.InfoLevel).Info("Adding deletion requested annotation to remote volume")
 						controller.AddAnnotation(remoteVolumeCopy, controller.DeletionRequested, "yes")
+
+						// also apply new annotation - SynchronizedDeletionStatus
+						log.V(common.InfoLevel).Info("Adding sync delete annotation to remote")
+						controller.AddAnnotation(remoteVolumeCopy, controller.SynchronizedDeletionStatus, "requested")
 						err := rClient.UpdatePersistentVolume(ctx, remoteVolumeCopy)
 						if err != nil {
+							log.V(common.InfoLevel).Info("Error encountered in updating remote volume")
 							return ctrl.Result{}, err
 						}
+
 						// Resetting the rate-limiter to requeue for the deletion of remote PV
 						return ctrl.Result{RequeueAfter: 1 * time.Millisecond}, nil
 					}
@@ -144,6 +151,8 @@ func (r *PersistentVolumeReconciler) Reconcile(ctx context.Context, req ctrl.Req
 				}
 			}
 		}
+
+		log.V(common.InfoLevel).Info("Removing finalizer on local volume")
 		finalizerRemoved := controller.RemoveFinalizerIfExists(volume, controller.ReplicationFinalizer)
 		if finalizerRemoved {
 			return ctrl.Result{}, r.Update(ctx, volume)
@@ -154,13 +163,22 @@ func (r *PersistentVolumeReconciler) Reconcile(ctx context.Context, req ctrl.Req
 
 	// Check for the finalizer; add, if doesn't exist
 	if finalizerAdded := controller.AddFinalizerIfNotExist(volumeCopy, controller.ReplicationFinalizer); finalizerAdded {
-		log.V(common.DebugLevel).Info("Finalizer not found adding it")
+		log.V(common.DebugLevel).Info("Finalizer not found, adding it")
 		return ctrl.Result{}, r.Update(ctx, volumeCopy)
 	}
 	// Check for deletion request annotation
 	if _, ok := volumeCopy.Annotations[controller.DeletionRequested]; ok {
-		log.V(common.InfoLevel).Info("Deletion Requested, annotation found and deleting the remote volume")
-		return ctrl.Result{}, r.Delete(ctx, volumeCopy)
+		log.V(common.InfoLevel).Info("Deletion Requested by remote's controller, annotation found")
+
+		// Check for the sync deletion annotation. If it's 'complete' or nonexistent, delete the PV.
+		// If it's 'requested', wait.
+		syncDeleteStatus, ok := volume.Annotations[controller.SynchronizedDeletionStatus]
+		if !ok || syncDeleteStatus == "complete" {
+			log.V(common.InfoLevel).Info("Synchronized Deletion annotation either complete or not set, deleting PV")
+			return ctrl.Result{}, r.Delete(ctx, volumeCopy)
+		} else {
+			log.V(common.InfoLevel).Info("Synchronized Deletion annotation exists and is not complete, cannot delete PV")
+		}
 	}
 
 	_, ok = volume.Annotations[controller.PVProtectionComplete]
@@ -303,6 +321,7 @@ func (r *PersistentVolumeReconciler) Reconcile(ctx context.Context, req ctrl.Req
 			log.V(common.DebugLevel).Info("Remote RG annotation has not been set on the local RG")
 		} else {
 			// Update the annotation for the remote PV object
+			// TODO: Also add local volume information to remote PV
 			controller.AddAnnotation(pv, controller.ReplicationGroup, remoteRGName)
 			controller.AddLabel(pv, controller.ReplicationGroup, remoteRGName)
 		}
